feat: Simplified pages by extracting components into separate files and  Removed redundant code and improved loading states for better user experience.
